+++ src/main/java/nimbus/command/CheckCommand.java
@@ -1,3 +1,12 @@
+package nimbus.command;
+
+import nimbus.exception.WrongDateTimeFormatException;
+import nimbus.task.DeadlineTask;
+import nimbus.task.EventTask;
+import nimbus.task.Task;
+import nimbus.ui.TaskList;
+import nimbus.ui.Ui;
+
 import java.time.LocalDate;
 import java.time.format.DateTimeFormatter;
 import java.time.format.DateTimeParseException;
