+++ src/main/java/nimbus/exception/MissingDeadlineException.java
@@ -1,0 +1,16 @@
+package nimbus.exception;
+
+import nimbus.ui.Ui;
+
+public class MissingDeadlineException extends Exception {
+
+
+    public MissingDeadlineException() {
+        super("Nimbus.Nimbus noticed that you did not include a deadline! Use /by!" + Ui.horizontalLine);
+    }
+
+    @Override
+    public String toString() {
+        return this.getMessage();
+    }
+}
