+++ src/main/java/nimbus/exception/MissingDescriptionException.java
@@ -1,8 +1,11 @@
+package nimbus.exception;
+
+import nimbus.ui.Ui;
+
 public class MissingDescriptionException extends Exception {
-    private static String horizontalLine = "\n-------------------------------------------------";
     public MissingDescriptionException(String taskName) {
         super("Oh noo, you cant leave the description of " + taskName +
-                " empty!!" + horizontalLine);
+                " empty!!" + Ui.horizontalLine);
     }
 
     @Override
