+++ src/main/java/nimbus/exception/WrongDateTimeFormatException.java
@@ -1,6 +1,8 @@
+package nimbus.exception;
+
 public class WrongDateTimeFormatException extends Exception {
     public WrongDateTimeFormatException() {
-        super("Wrong DateTime format, Follow DD/MM/YYYY 24HR Time \n" +
+        super("Wrong nimbus.ui.DateTime format, Follow DD/MM/YYYY 24HR Time \n" +
                 "Eg: 2/12/2019 1800");
     }
 }
