+++ src/main/java/nimbus/task/Task.java
@@ -1,3 +1,5 @@
+package nimbus.task;
+
 public class Task {
     private String taskName;
     private boolean completed;
