+++ src/main/java/nimbus/task/TodoTask.java
@@ -1,3 +1,5 @@
+package nimbus.task;
+
 public class TodoTask extends Task{
     private String taskName;
